Divisione funzioni e cambio barra di progressione (codice finale)
